Move LightblueAuditLogWritter into its own file
