+++ common/src/main/java/com/redhat/lightblue/rest/audit/LightblueAuditServletFilter.java
@@ -4,13 +4,10 @@ import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.security.Principal;
-import java.text.SimpleDateFormat;
-import java.util.Date;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.RejectedExecutionException;
 import java.util.concurrent.TimeUnit;
-import java.util.regex.Matcher;
 
 import javax.servlet.Filter;
 import javax.servlet.FilterChain;
@@ -28,7 +25,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.common.base.Stopwatch;
-import com.redhat.lightblue.rest.audit.LightblueOperationChecker.Info;
 
 /**
  * Filter all the request which must have data or metadata as their context path
@@ -39,9 +35,7 @@ import com.redhat.lightblue.rest.audit.LightblueOperationChecker.Info;
 // Handle any request
 public class LightblueAuditServletFilter implements Filter {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(LightblueAuditServletFilter.class);
-
-    private static final String YYYY_MM_DD_T_HH_MM_SS_SSSZ = "yyyy-MM-dd'T'HH:mm:ss.SSSZ";
+    static final Logger LOGGER = LoggerFactory.getLogger(LightblueAuditServletFilter.class);
 
     private static final ExecutorService jobExecutor = Executors.newCachedThreadPool();
 
@@ -87,7 +81,6 @@ public class LightblueAuditServletFilter implements Filter {
             try {
                 logEntryBuilder.setPrincipal(p);
                 logEntryBuilder.setRequestSize(hReq.getContentLength());
-                logEntryBuilder.setTimestampText(new SimpleDateFormat(YYYY_MM_DD_T_HH_MM_SS_SSSZ).format(new Date()));
                 logEntryBuilder.setResource(hReq.getContextPath());
 
                 if (res instanceof HttpServletResponse) {
@@ -114,7 +107,7 @@ public class LightblueAuditServletFilter implements Filter {
                 }
 
                 //Log Async. No reason to hold up the response for auditing purposes.
-                jobExecutor.execute(new AuditLogWritter(logEntryBuilder, hReq, res, isMetadata));
+                jobExecutor.execute(new LightblueAuditLogWritter(logEntryBuilder, hReq, res, isMetadata));
 
             } catch (RejectedExecutionException e) {
                 LOGGER.warn("Audit thread rejected from executor", e);
@@ -139,8 +132,8 @@ public class LightblueAuditServletFilter implements Filter {
         LOGGER.debug("Initializing LightblueAuditServletFilter");
     }
 
-    private static class HttpServletResponseWrapperBuffered extends HttpServletResponseWrapper {
-        private final ByteArrayPrintWriter byteArrayPrintWriter;
+    static class HttpServletResponseWrapperBuffered extends HttpServletResponseWrapper {
+        final ByteArrayPrintWriter byteArrayPrintWriter;
 
         public HttpServletResponseWrapperBuffered(HttpServletResponse httpServletResponse, ByteArrayPrintWriter byteArrayPrintWriter) {
             super(httpServletResponse);
@@ -159,9 +152,9 @@ public class LightblueAuditServletFilter implements Filter {
 
     }
 
-    private static class ByteArrayPrintWriter {
+    static class ByteArrayPrintWriter {
         private final ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
-        private final PrintWriter printWriter = new PrintWriter(byteArrayOutputStream);
+        final PrintWriter printWriter = new PrintWriter(byteArrayOutputStream);
         private final ServletOutputStream byteArrayServletStream = new ByteArrayServletStream(byteArrayOutputStream);
 
         public PrintWriter getWriter() {
@@ -188,209 +181,5 @@ public class LightblueAuditServletFilter implements Filter {
         public void write(int param) throws IOException {
             byteArrayOutputStream.write(param);
         }
-    }
-
-    /**
-     * Allows expensive audit work to be performed asynchronously from the actual user request.
-     *
-     * @author dcrissman
-     */
-    private static class AuditLogWritter implements Runnable {
-
-        private final LogEntryBuilder logEntryBuilder;
-        private final boolean isMetadata;
-        private final HttpServletRequest req;
-        private final ServletResponse res;
-
-        public AuditLogWritter(LogEntryBuilder logEntryBuilder, HttpServletRequest req, ServletResponse res, boolean isMetadata) {
-            this.logEntryBuilder = logEntryBuilder;
-            this.isMetadata = isMetadata;
-            this.req = req;
-            this.res = res;
-        }
-
-        @Override
-        public void run() {
-            Info info = parseOperationEntityVersionStatus(req, isMetadata, logEntryBuilder);
-            if (info != null) {
-                logEntryBuilder.setOperation(info.operation);
-                logEntryBuilder.setEntityName(info.entity);
-                logEntryBuilder.setEntityVersion(info.version);
-                logEntryBuilder.setEntityStatus(info.status);
-            }
-
-            if (res instanceof HttpServletResponseWrapperBuffered) {
-                HttpServletResponseWrapperBuffered httpServletResponseWrapperBuffered = (HttpServletResponseWrapperBuffered) res;
-                httpServletResponseWrapperBuffered.byteArrayPrintWriter.printWriter.flush();
-                byte[] bytes = httpServletResponseWrapperBuffered.byteArrayPrintWriter.toByteArray();
-                try {
-                    res.getOutputStream().write(bytes);
-                    logEntryBuilder.setResponseSize(bytes.length);
-                } catch (IOException e) {
-                    LOGGER.warn("Unable to write response bytes for audit", e);
-                }
-            }
-
-            final LogEntry logEntry = logEntryBuilder.createLogEntry();
-            String logEntryString = String.format(
-                    "Audited lightblue rest request => " +
-                            "{ " +
-                            "\"initialTimestamp\":\"%s\", " +
-                            "\"currentTimestamp\":\"%s\" , " +
-                            "\"principal\":\"%s\" , " +
-                            "\"resource\":\"%s\" , " +
-                            "\"operation\":\"%s\" , " +
-                            "\"entityName\":\"%s\" , " +
-                            "\"entityVersion\":\"%s\" , " +
-                            "\"entityStatus\":\"%s\" , " +
-                            "\"requestSize\":\"%d\" , " +
-                            "\"responseSize\":\"%d\" , " +
-                            "\"timeElapsedInNs\":\"%d\"  " +
-                            " }",
-                    logEntry.getTimestampText(),
-                    new SimpleDateFormat(YYYY_MM_DD_T_HH_MM_SS_SSSZ).format(new Date()),
-                    logEntry.getPrincipal().getName(),
-                    logEntry.getResource(),
-                    logEntry.getOperation(),
-                    logEntry.getEntityName(),
-                    logEntry.getEntityVersion(),
-                    logEntry.getEntityStatus(),
-                    logEntry.getRequestSize(),
-                    logEntry.getResponseSize(),
-                    logEntry.getTimeElapsedInNs()
-                    );
-
-            LOGGER.info(logEntryString);
-        }
-
-        private LightblueOperationChecker.Info parseOperationEntityVersionStatus(HttpServletRequest hReq, boolean isMetadata, LogEntryBuilder logEntryBuilder) {
-            // List of methods in http://www.w3.org/Protocols/HTTP/Methods.html
-            String method = hReq.getMethod().toUpperCase();
-            boolean processed = false;
-            String servletPath = hReq.getServletPath();
-
-            LightblueOperationChecker.Info info = null;
-            Matcher m = null;
-            switch (method) {
-                case "GET":
-                    if (!processed) {
-                        processed = true;
-                        if (!isMetadata) {
-                            if ((info = LightblueCrudOperationChecker.simpleFindVersionRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueCrudOperationChecker.simpleFindRegex.matches(servletPath)).found) {
-                                return info;
-                            }
-                        } else {
-                            if ((info = LightblueMetadataOperationChecker.getDepGraphVersionRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueMetadataOperationChecker.getDepGraphEntityRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueMetadataOperationChecker.getDepGraphRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueMetadataOperationChecker.getEntityRolesVersionRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueMetadataOperationChecker.getEntityRolesEntityRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueMetadataOperationChecker.getEntityRolesRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueMetadataOperationChecker.getEntityNamesRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueMetadataOperationChecker.getEntityNamesStatusRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueMetadataOperationChecker.getEntityVersionsRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueMetadataOperationChecker.getMetadataRegex.matches(servletPath)).found) {
-                                return info;
-                            }
-                        }
-                    }
-                case "POST":
-                    if (!processed) {
-                        processed = true;
-                        if (!isMetadata) {
-                            if ((info = LightblueCrudOperationChecker.findRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueCrudOperationChecker.findVersionRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueCrudOperationChecker.deleteRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueCrudOperationChecker.deleteVersionRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueCrudOperationChecker.updateRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueCrudOperationChecker.updateVersionRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueCrudOperationChecker.saveRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueCrudOperationChecker.saveVersionRegex.matches(servletPath)).found) {
-                                return info;
-                            }
-
-                        } else {
-                            if ((info = LightblueMetadataOperationChecker.setDefaultVersionRegex.matches(servletPath)).found) {
-                                return info;
-                            }
-                        }
-                    }
-                case "PUT":
-                    if (!processed) {
-                        processed = true;
-                        if (!isMetadata) {
-                            if ((info = LightblueCrudOperationChecker.insertRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueCrudOperationChecker.insertVersionRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueCrudOperationChecker.insertAltRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueCrudOperationChecker.insertAltVersionRegex.matches(servletPath)).found) {
-                                return info;
-                            }
-                        } else {
-                            if ((info = LightblueMetadataOperationChecker.createSchemaRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueMetadataOperationChecker.updateSchemaStatusRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueMetadataOperationChecker.updateEntityInfoRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueMetadataOperationChecker.createMetadataRegex.matches(servletPath)).found) {
-                                return info;
-                            }
-                        }
-                    }
-                case "DELETE":
-                    if (!processed) {
-                        if (isMetadata) {
-                            processed = true;
-                            if ((info = LightblueMetadataOperationChecker.clearDefaultVersionRegex.matches(servletPath)).found) {
-                                return info;
-                            } else if ((info = LightblueMetadataOperationChecker.removeEntityRegex.matches(servletPath)).found) {
-                                return info;
-                            }
-                        }
-                    }
-                case "HEAD":
-                case "CONNECT":
-                case "OPTIONS":
-                case "TRACE":
-                case "SEARCH":
-                case "SPACEJUMP":
-                case "CHECKIN":
-                case "UNLINK":
-                case "SHOWMETHOD":
-                case "CHECKOUT":
-                    // Valid HTTP method but not used yet
-                default:
-                    if (processed) {
-                        // The URL missed all the patterns. Maybe there is a not mapped rest service or the URL is invalid
-                        LOGGER.warn("The URL doesn't map to one of the rest services. Request URI: " + hReq.getRequestURI()); // TODO Unique case where the whole URI is logged, when it is an exception
-                    } else {
-                        // Called on of the not mapped HTTP methods
-                        LOGGER.info("Invalid HTTP method: " + method);
-                    }
-            }
-            return null;
-        }
-
     }
 }
